# docs/openapi/v2.yaml
openapi: 3.1.0
info:
  title: Bazaar v2 API
  version: 2.0.0
  description: |
    Unified BFF API for the Bazaar platform.  All endpoints live under `/v2/*`.

    **Platform routing** — every request must carry an `X-Client` header:
    - `mobile` — Flutter app (bearer tokens).
    - `web` — Next.js storefront (cookies preferred).

    Endpoints marked **mobile-only** or **web-only** return `400 UNKNOWN_PLATFORM`
    when called from the wrong client.  Endpoints under the **Shared** group
    (Products, Wishlist) are served to both clients.

    **Response envelope** — every response, success or error, is wrapped:

    ```json
    { "success": true, "data": { … } }
    { "success": false, "error": { "code": "TOKEN_EXPIRED", "message": "…" } }
    ```

    Paginated responses add a `meta` key alongside `data`.

    **Auth** — two complementary schemes:
    - `bearerAuth` — `Authorization: Bearer <access-jwt>` (mobile default, web fallback).
    - `cookieAuth` — `bz_at` HttpOnly cookie (web default).

    The app accepts either; prefer what matches `tokenDelivery` at login.

    **Error codes** — machine-readable codes in `error.code`.  Full catalog:
    `docs/v2-auth/ERROR-CATALOG.md`.
  contact:
    name: Bazaar Platform Team
    email: platform@bazaar-uae.com
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: "${BACKEND_URL}/v2"
    description: Production (set BACKEND_URL in .env)
  - url: "${STAGING_URL}/v2"
...

...
# Tags
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access JWT obtained from `/v2/auth/login` (or OAuth login) with
        `tokenDelivery: "bearer"`.  Pass as `Authorization: Bearer <token>`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: bz_at
      description: >
        HttpOnly access-token cookie set by the server when
        `tokenDelivery: "cookie"` is used.  Web clients use this by default.

# ─────────────────────────────────────────────────────────────────────────────
# Reusable Schemas
# ─────────────────────────────────────────────────────────────────────────────
  schemas:
...
    Envelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          description: Payload. Present on success; null or absent on error.
        error:
          $ref: '#/components/schemas/ErrorBody'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: TOKEN_EXPIRED
          description: >
            Machine-readable stable identifier. Clients switch on this value.
            Full catalog in `docs/v2-auth/ERROR-CATALOG.md`.
        message:
          type: string
          example: Token expired.
          description: Human-readable English message. Do not parse; use `code`.
        details:
          description: Optional extra context — validation field errors, etc.
          nullable: true
    PaginationMeta:
      type: object
      required: [total, page, limit, pages]
      properties:
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        pages:
          type: integer
          example: 3

    # ── User / Session ───────────────────────────────────────────────────────
...
  responses:
...
    401Unauthorized:
      description: Authentication required or token invalid/expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          example:
            success: false
            error:
              code: TOKEN_EXPIRED
              message: Token expired.
...
    500Internal:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: Something went wrong. Please try again.

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /payment-methods:
    get:
      operationId: listPaymentMethods
      tags: [Orders]
      summary: List available payment methods (mobile)
      description: |
        Returns the payment methods available to the current user.
        **Mobile-only**.

        > **NOTE** — the `enabled` flag on each payment method is a runtime-toggleable value,
        > not a static constant.  Ops can flip individual providers on/off without a redeploy
        > via `PUT /v2/admin/payment-method-config`.  Env-var hard gates (e.g. `TABBY_AUTH_KEY`)
        > still control whether a provider appears in the list at all.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XClientMobile'
      responses:
        '200':
          description: Payment methods list.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            name:
                              type: string
                            enabled:
                              type: boolean
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '500':
          $ref: '#/components/responses/500Internal'
...